backend/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
python-multipart==0.0.20
pillow==12.0.0
httpx==0.28.1
python-dotenv==1.2.1
pydantic-settings==2.6.1
redis==5.0.8
pytest==9.0.1
pytest-asyncio==1.3.0
black==24.8.0
ruff==0.6.8
